src/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.101.0
requests==2.31.0
pywin32==311
